WIP: totally draft progress to test rure-go (Rust regex library)
It's definitely a speed machine for many cases! Below are a couple of
the existing benchmarks I have which use regexes.

The tt.10 script just uses the trivial regex /.$/ which I believe is
the kind of thing rure actually has fast paths for (I guess in this
case it's saying "is there at least one character?"). But amazing to
see how much gain there is to be had from this kind of fast path.

The tt.15 script uses a bunch of more complicated regexes, as well as
other AWK logic. From a quick profile it looks like it's measuring
mostly the speed of the virtual machine stack operations and (GoAWK's
relatively poor) string concatenation.

Test      | goawk |  orig |   awk |  gawk |  mawk
--------- | ----- | ----- | ----- | ----- | -----
tt.10     |  2.66 |  0.43 |  1.00 |  1.59 |  3.70  # 6.2 times as fast!
tt.15     |  0.74 |  0.70 |  1.00 |  1.00 |  2.20  # only 6% faster
--------- | ----- | ----- | ----- | ----- | -----

And here's a custom one-liner, with a more complex regex of the kind
you'd use for data processing. In this case it's finding the number of
lines the word Benjamin or Menno appears in 10 concatenated copies of
the King James Bible (Menno gave me this idea -- that name doesn't
appear).

$ time ./orig '/\b(Benjamin|Menno)\b/ { n++ } END { print n }' <../countwords/kjvbible_x10.txt
1660
real 0m1.751s
$ time ./goawk '/\b(Benjamin|Menno)\b/ { n++ } END { print n }' <../countwords/kjvbible_x10.txt
1660
real 0m0.373s  # 4.7 times as fast!

I'm not going to us this for the main version of GoAWK, as it's a heavy
dependency that's a bit of a pain because it relies on a Rust shared
library and CGo, but it might be worth maintaining a proper branch.